Refactor Link component for improved maintainability and accessibility
I've refactored and move the Link component in src/components/Link.tsx to src/components/atoms/links/Link.tsx to improve maintainability and accessibility.

The changes include:
- Setting the isNewTab prop default to false.
- Moving the onClick and onKeyUp handlers to the anchor tag for better accessibility and HTML semantics.
- Removing unnecessary conversion of the href prop to a string.
- Simplifying the logic for handling the trackClick prop.
- Adding e.preventDefault() to the handleClick and handleKeyUp event handlers to prevent the default action of the anchor tag when the trackClick prop is provided.

Additionally, I have modified the component's export statement to only export the default Link component, following the project's pattern.
