+++ src/apps/intermediate-list/modal/my-payment-overview-modal.tsx
@@ -2,7 +2,7 @@ import React, { FC } from "react";
 import ExternalLinkIcon from "@danskernesdigitalebibliotek/dpl-design-system/build/icons/buttons/icon-btn-external-link.svg";
 import Modal, { useModalButtonHandler } from "../../../core/utils/modal";
 import { useText } from "../../../core/utils/text";
-import { Link } from "../../../components/atoms/link";
+import Link from "../../../components/atoms/links/Link";
 import { useUrls } from "../../../core/utils/url";
 
 const MyPaymentOverviewModal: FC = () => {
