+++ src/components/search-result-list/search-result-list-item/search-result-list-item.tsx
@@ -8,7 +8,7 @@ import ButtonFavourite, {
   ButtonFavouriteId
 } from "../../button-favourite/button-favourite";
 import { CoverProps } from "../../cover/cover";
-import { Link } from "../../atoms/link";
+import Link from "../../atoms/links/Link";
 import {
   creatorsToString,
   filterCreators,
